Rename backend directory to api
